fix-update-services
